Refactor syntax highlighting into SRP modules
