+++ runtime/src/query/highlights/capture_names.rs
@@ -1,0 +1,52 @@
+//! Standard Tree-sitter capture names used for syntax highlighting.
+
+/// Comment capture.
+pub const COMMENT: &str = "comment";
+/// String literal capture.
+pub const STRING: &str = "string";
+/// Numeric literal capture.
+pub const NUMBER: &str = "number";
+/// Language keyword capture.
+pub const KEYWORD: &str = "keyword";
+/// Function definition capture.
+pub const FUNCTION: &str = "function";
+/// Function call capture.
+pub const FUNCTION_CALL: &str = "function.call";
+/// Method definition capture.
+pub const METHOD: &str = "method";
+/// Method call capture.
+pub const METHOD_CALL: &str = "method.call";
+/// Variable capture.
+pub const VARIABLE: &str = "variable";
+/// Built-in variable capture.
+pub const VARIABLE_BUILTIN: &str = "variable.builtin";
+/// Constant capture.
+pub const CONSTANT: &str = "constant";
+/// Built-in constant capture.
+pub const CONSTANT_BUILTIN: &str = "constant.builtin";
+/// Type capture.
+pub const TYPE: &str = "type";
+/// Built-in type capture.
+pub const TYPE_BUILTIN: &str = "type.builtin";
+/// Property capture.
+pub const PROPERTY: &str = "property";
+/// Operator capture.
+pub const OPERATOR: &str = "operator";
+/// Bracket punctuation capture.
+pub const PUNCTUATION_BRACKET: &str = "punctuation.bracket";
+/// Delimiter punctuation capture.
+pub const PUNCTUATION_DELIMITER: &str = "punctuation.delimiter";
+/// Special punctuation capture.
+pub const PUNCTUATION_SPECIAL: &str = "punctuation.special";
+/// Attribute capture.
+pub const ATTRIBUTE: &str = "attribute";
+/// Namespace capture.
+pub const NAMESPACE: &str = "namespace";
+/// Module capture.
+pub const MODULE: &str = "module";
+/// Label capture.
+pub const LABEL: &str = "label";
+/// Tag capture.
+pub const TAG: &str = "tag";
+/// Syntax error capture.
+pub const ERROR: &str = "error";
